Add files via upload
Added the general functionality I was looking for when I started writing this. The CSV file contains all of the books that could be presented to the players (and adding new books to the list is straightforward). The Book object contains all information about that book, and for readability the libraryObject class contains a collection of these books.